+++ tests/test_app/models.py
@@ -1,3 +1,5 @@
+import uuid
+
 from django.db import models
 
 
@@ -9,3 +11,11 @@ class Item(models.Model):
 
     class Meta:
         ordering = ["name"]
+
+
+class ItemWithUUID(models.Model):
+    id = models.UUIDField(primary_key=True, default=uuid.uuid4)
+    name = models.CharField(max_length=50)
+
+    def __str__(self):
+        return f"{self.name}({self.id})"